v1.3 - small updates and removed NBT file storage on stop
Command now gives feedback (success and error messages) to user in most, if not all use cases. Also removed the system which stored all UUIDs in an NBT file on stop, and loaded them to variables on start. Seeing as the UUIDs are being stored anyway, this system, to my knowledge, was completely redundant.