v1.5 - remove warnings and update code
Changed the lines where the text displays are spawned to not use SkBee's effect anymore, as Skript added support for this a while back. This might mean that this version doesn't work on Skript 2.6.4 or older. Also just freshened up the rest of the code to be more like my recent scripts.